project/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
asyncpg==0.29.0
kafka-python-ng==2.2.2
pydantic-settings==2.1.0
httpx==0.26.0
textblob==0.17.1
